+++ src/constants.tsx
@@ -16,6 +16,9 @@ export const ROUTES: RouteDefinition[] = [
   { path: '/game', component: lazy(() => import('./pages/game')) },
   { path: '/game/:id', component: lazy(() => import('./pages/game/[id]')) },
   { path: '/game/:id/play', component: lazy(() => import('./pages/game/[id]/play')) },
+  { path: '/game/:id/single', component: lazy(() => import('./pages/game/[id]/single')) },
+  { path: '/game/:id/multi', component: lazy(() => import('./pages/game/[id]/multi')) },
+  { path: '/game/:id/custom', component: lazy(() => import('./pages/game/[id]/custom')) },
   { path: '/tape', component: lazy(() => import('./pages/tape')) },
   { path: '/bot', component: lazy(() => import('./pages/bot')) },
   { path: '/rate', component: lazy(() => import('./pages/rate')) },
