--- src/pages/game/[id]/components/DebugMode/comp.tsx
@@ -1,5 +1,0 @@
-export const DebugMode = () => {
-  return (
-    <h1>Debug Mode</h1>
-  );
-};
