--- src/pages/game/[id]/components/MultiMode/components/CustomMode/index.ts
@@ -1,1 +1,0 @@
-export * from './comp';
