--- src/pages/game/[id]/components/MultiMode/components/MatchMode/components/RoomComp/index.ts
@@ -1,1 +1,0 @@
-export * from './comp';
