--- src/pages/game/[id]/components/MultiMode/components/MatchMode/constants.ts
@@ -1,12 +1,0 @@
-import { Room } from './types';
-
-export const DEBUG_ROOM: Room = {
-  roomId: 'KHRLJBWPOklwjefklnvL',
-  players: [{
-    id: 'andrew',
-    score: 1500,
-  }, {
-    id: 'andrewleung',
-    score: 1480,
-  }],
-};
