--- src/pages/game/[id]/components/MultiMode/components/MatchMode/index.ts
@@ -1,1 +1,0 @@
-export * from './comp';
