--- src/pages/game/[id]/components/MultiMode/components/index.ts
@@ -1,2 +1,0 @@
-export * from './MatchMode';
-export * from './CustomMode';
