--- src/pages/game/[id]/components/SingleMode/index.ts
@@ -1,1 +1,0 @@
-export * from './comp';
