--- src/pages/game/[id]/components/index.ts
@@ -1,3 +1,0 @@
-export * from './MultiMode';
-export * from './DebugMode';
-export * from './SingleMode';
