+++ src/pages/game/[id]/multi/components/index.ts
