+++ src/utils/createEvent.ts
@@ -1,8 +1,9 @@
 import { Socket } from 'socket.io-client';
 import { Accessor, createEffect } from 'solid-js';
+import { isString } from 'lodash-es';
 
-export const createEvent = (socket: Accessor<Socket | undefined>, event: string, callback: (...args: any[]) => void) => {
+export const createEvent = (socket: Accessor<Socket | undefined>, event: string | Accessor<string>, callback: (...args: any[]) => void) => {
   createEffect(() => {
-    socket()?.on(event, callback);
+    socket()?.on(isString(event) ? event : event(), callback);
   });
 };
